refactor(backend): moved /files to separate router
